# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Livestream Studio API - OBS-like web application
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/export:
    get:
      operationId: exportProject
      tags: [projects]
      summary: Export project as JSON
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project JSON export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExport"
        "404":
          description: Not found
  /projects/import:
    post:
      operationId: importProject
      tags: [projects]
      summary: Import project from JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExport"
      responses:
        "201":
          description: Imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ── Scenes ────────────────────────────────────────────────
  /projects/{projectId}/scenes:
    get:
      operationId: listScenes
      tags: [scenes]
      summary: List all scenes in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
    post:
      operationId: createScene
      tags: [scenes]
      summary: Create a new scene
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
  /scenes/{id}:
    patch:
      operationId: updateScene
      tags: [scenes]
      summary: Update a scene
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "404":
          description: Not found
    delete:
      operationId: deleteScene
      tags: [scenes]
      summary: Delete a scene
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scenes/{id}/duplicate:
    post:
      operationId: duplicateScene
      tags: [scenes]
      summary: Duplicate a scene
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "404":
          description: Not found
  /scenes/reorder:
    post:
      operationId: reorderScenes
      tags: [scenes]
      summary: Reorder scenes within a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneReorder"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"

  # ── Sources ───────────────────────────────────────────────
...
  /audio/tracks:
    get:
      operationId: listAudioTracks
      tags: [audio]
      summary: List all audio tracks / mixer channels
      responses:
        "200":
          description: List of audio tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AudioTrack"
    post:
      operationId: createAudioTrack
      tags: [audio]
      summary: Create an audio track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioTrackInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTrack"
  /audio/tracks/{id}:
    patch:
      operationId: updateAudioTrack
      tags: [audio]
      summary: Update audio track settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioTrackUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTrack"
        "404":
          description: Not found
    delete:
      operationId: deleteAudioTrack
      tags: [audio]
      summary: Delete audio track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Uploads ───────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        activeSceneId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        activeSceneId:
          type: ["integer", "null"]
    ProjectExport:
      type: object
      required: [name, scenes]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
        outputConfig:
          $ref: "#/components/schemas/OutputConfig"
        streamConfig:
          $ref: "#/components/schemas/StreamConfig"

    # ── Scene ─────────────────────────────────────────────────
    Scene:
      type: object
      required: [id, projectId, name, sortOrder, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        thumbnail:
          type: ["string", "null"]
        transitionType:
          type: string
          enum: [cut, fade, slide, swipe, zoom, dissolve]
          default: fade
        transitionDurationMs:
          type: integer
          default: 300
        sortOrder:
          type: integer
        createdAt:
          type: string
    SceneInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        transitionType:
          type: string
          enum: [cut, fade, slide, swipe, zoom, dissolve]
        transitionDurationMs:
          type: integer
    SceneUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        transitionType:
          type: string
          enum: [cut, fade, slide, swipe, zoom, dissolve]
        transitionDurationMs:
          type: integer
        sortOrder:
          type: integer
    SceneReorder:
      type: object
      required: [projectId, sceneIds]
      properties:
        projectId:
          type: integer
        sceneIds:
          type: array
          items:
            type: integer

    # ── Source ────────────────────────────────────────────────
...
    StreamConfig:
      type: object
      required: [id, rtmpUrl, streamKey]
      properties:
        id:
          type: integer
        rtmpUrl:
          type: string
        streamKey:
          type: string
        platform:
          type: string
          enum: [facebook, youtube, twitch, custom]
...
    OutputConfig:
      type: object
      required: [id, resolution, fps, videoBitrate, audioBitrate, encoder]
      properties:
        id:
          type: integer
        resolution:
          type: string
          enum: ["720p", "1080p", "1440p", "4K"]
        fps:
          type: integer
          enum: [24, 30, 60]
        videoBitrate:
          type: integer
        audioBitrate:
          type: integer
        encoder:
          type: string
          enum: [H264, H265, VP9, AV1]
        recordingEnabled:
          type: boolean
        recordingFormat:
          type: string
          enum: [mp4, mkv, mov]
...
    AudioTrack:
      type: object
      required: [id, name, gain, volume, muted, solo]
      properties:
        id:
          type: integer
        name:
          type: string
        sourceId:
          type: ["integer", "null"]
        gain:
          type: number
        volume:
          type: number
        balance:
          type: number
        muted:
          type: boolean
        solo:
          type: boolean
        filters:
          type: object
          additionalProperties: true
    AudioTrackInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        sourceId:
          type: integer
        gain:
          type: number
        volume:
          type: number
        balance:
          type: number
    AudioTrackUpdate:
      type: object
      properties:
        name:
          type: string
        gain:
          type: number
        volume:
          type: number
        balance:
          type: number
        muted:
          type: boolean
        solo:
          type: boolean
        filters:
          type: object
          additionalProperties: true

    # ── Media Assets ──────────────────────────────────────────
...