# lib/api-zod/src/generated/types/audioTrackInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Livestream Studio API - OBS-like web application
 * OpenAPI spec version: 0.1.0
 */

export interface AudioTrackInput {
  /** @minLength 1 */
  name: string;
  sourceId?: number;
  gain?: number;
  volume?: number;
  balance?: number;
}